refactor: 헬퍼 함수, 타입들 유틸리티 폴더로 이동
