--- src/components/RecruitForm/type.ts
@@ -1,12 +1,0 @@
-export interface RecruitFormValues {
-  category: string;
-  participants: {
-    study: Participant[];
-    project: Participant[];
-  };
-}
-
-type Participant = {
-  part: string;
-  count: number;
-};
