+++ src/components/RecruitForm/utils/constants.ts
@@ -1,0 +1,5 @@
+export const fieldNames = {} as const;
+
+// meta
+export const parts = ['프론트엔드', '백엔드', '기획/디자인', '앱'];
+export const pf = [];
