+++ src/components/RecruitForm/utils/type.ts
@@ -1,0 +1,12 @@
+export interface RecruitFormValues {
+    category: string;
+    participants: {
+        study: Participant[];
+        project: Participant[];
+    };
+}
+
+export type Participant = {
+    part: string;
+    count: number;
+};
